# swagger/something.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "storage",
    "version": "0.0.1"
  },
  "host": ".endpoints.phdigidev.cloud.goog",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/iteratorby": {
      "post": {
        "operationId": "IteratorBy",
        "responses": {
          "200": {
            "description": "(streaming responses)",
            "schema": {
              "$ref": "#/definitions/IteratorByRes"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IteratorByReq"
            }
          }
        ],
        "tags": [
          "Storage"
        ]
      }
    },
    "/v1/multi/get": {
      "post": {
        "operationId": "GetMulti",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetMultiRes"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetMultiReq"
            }
          }
        ],
        "tags": [
          "Storage"
        ]
      }
    },
    "/v1/multi/set": {
      "post": {
        "operationId": "SetMulti",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/SetMultiRes"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SetMultiReq"
            }
          }
        ],
        "tags": [
          "Storage"
        ]
      }
    },
...
  },
  "definitions": {
...
    "GetMultiReq": {
      "type": "object",
      "properties": {
        "IDs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "title": "GetMulti"
    },
    "GetMultiRes": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "format": "int64"
        },
        "description": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        }
      }
    },
...
    "Item": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        },
        "timestamp": {
          "type": "string",
          "format": "int64"
        },
        "keys": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "byte"
          }
        },
        "deleted": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "IteratorByReq": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "op": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      },
      "title": "IteratorBy - this should reeeeaaallly be a stream"
    },
    "IteratorByRes": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "format": "int64"
        },
        "description": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        }
      }
    },
...
    "SetMultiReq": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        }
      },
      "title": "SetMulti"
    },
    "SetMultiRes": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "format": "int64"
        },
        "description": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "firebase": {
      "type": "oauth2",
      "description": "Used to authenticate Firebase",
      "in": "header",
      "flow": "implicit",
      "authorizationUrl": "https://securetoken.google.com/phdigidev"
    }
  },
  "security": [
    {
      "firebase": []
    }
  ]
}